Make alphaTest mask consistent for all cases and use the correct value
The code used a mix of "test high bit" and "test for zero value" for
the alphaTest blit mask. Replace the constant with a symbolic version
and ensure that the test is correctly applied to "zero value" case, not
all cases where the high bit is set.

Partial fix for issue #629