Fix bug #645: Tabs cause rendered line overflow in block formatting
Add a GS_FIXED glyph flag to mark that a glyph should not be moved
during justification.

Justify the part of the line after the last GS_FIXED glyph, or from
the start of the line if there is no GS_FIXED glyph.

If GS_ISFIRST is set in nextflags, don't set GS_FIXED in it.