eTextPara: Fix bounding box weirdness
Setting the left bound of a glyph was originally (spacing mine):

    ng.bbox.setLeft( (flags & GS_ISFIRST | glyphs.empty()
                      ? cursor.x() : cursor.x()-1) + glyph->left );

Presumably this means "for the first glyph on a line use x, otherwise
one less".  Why one less?  Over the years it has become (spacing mine):

    ng.bbox.setLeft(((flags & GS_ISFIRST) | cursor.x())
                    + left + xborder);

So it's dropped the "one less", but now binary ors the flag and cursor.
Since it makes no sense remove the flag test altogether.