Only strip bouquet line newline if it has one
Avoid stripping the last character of an input line in
eDVBDB::loadBouquet() if the character is not a newline.

Stripping of the return character has been left in to allow backward
compatibility for old bouquets (e.g. from old scans or settings
backups) that still have \r\n line endings.