servicedvb.cpp: Seek to 1 second from end of buffer when entering timeshift
Testing if this improves the reliability of timeshift. The playback
state machine can get confused if it can not correctly determine the PTS.
This appears to be more likely right at the end of the timeshift buffer.
This change is intended to test whether things improve. It is not a fix.
The PTS calculation needs to be made more robust and the state machine
should be more forgiving.

Thanks to Peter Urbanec