Recovery inetd.conf by default complete for IPv4/6 from IPv6Setup
With this change we can restore complete inetd.conf by default for both IPv4/6  from the IpV6Setup screen with blue button.

This depends on whether it is enabled/disabled and then we press directly blue...also as always can be done from the green button once it is enabled/disabled.

I have tried the nmdb and smdb service to put them enabled for both media, but this should continue as it is now not enabled since it consumes a lot of CPU.

added devicepic (pixmap):

I add icon to show that it is enabled, it shows ethernet icon for LAN and Wireless icon for WLAN.
Example: if we have LAN enabled, it shows an icon in LAN connection, if we enter WLAN it shows the same icon, which indicates that the current connection is LAN, it would show us the icon Not enabled wireless.

@fairbird you have to adapt this to the skin, something like this:

https://github.com/OpenVisionE2/TurquoiseHD-skin/commit/7ccdb68ee92ee7e684c3eaeebac27b55fdfefca3