Make the buttons move only on condition.
I always want an intuitive GUI and as simple as possible, I hate to show buttons when they still do not apply, the user can doubt the actions of the colored buttons ... it can be the beginning to make a more elegant code in skins .
It's simple, StaticText can be conditional, let's take advantage of this for our source. (view skin.templates.xml turquoise->screen line 357).