Push TMBD in Red Button SingleEPG/MultiEPG
I push red button in SingleEPG for TMBD plugin @ Dima73 and colaborations Taapat.

The zap function from the red button is removed from MultiEPG, now start the plugin.

@persianpross necessary so that the user does not have to install the plugin manually that compiles with that plugin.

In any case, if the user does not have it installed, he will be notified with messagebox that he has to do so if he wants to use said button, since the key of that button without said "TMBD Not Installed" plugin, once installed the button key will be "Search TMBD"