Fix for HDD format getting stuck at 51%
On HDDs that previously had DOS/NTFS partition in the exact same spot that
a new ext4 partition is being created, mkfs.ext4 would hang, asking if the
user wants to clobber the old partition. This would in turn result in the
disk format job being stuck at 51%.

Pass options to force the creation of an ext4 file system so that the
disk format can complete even on HDDs that are being reused from Windows
machines.