Update for Samba 4.10.17
-Updated log.

For now until we have a clean script start after installing this service, we do an init 6 to restart STB and ignore the callback that also ends with a TryQuitMainloop 2, with this we fix the output of a messagebox loop.

This is caused by a start of samba.sh in its first installation not clean, although later the service is handled perfectly in the setup, autostart, stop, start etc everything works properly ... in fact the sections of the STB filesystem in a windows PC works without problems.

Callback will resume as long as it perceives that we have a clean log in the samba installation.