PicCript.py: Fix Nagravisión encryption system detection
Nagra system detects none.png, since this system is different from self.CHANGED_CLEAR, that condition must be extended to be correctly detected in skin.
In addition, two digits for caid is not enough to differentiate other encryption systems, you have to broaden the search criteria to display them correctly in skin.
To show FTA channel we already have PicEmu2