[config.py] Rename KEY_ values to ACTIONKEY_
In config.py the values KEY_xxx are *NOT* the same as the remote control button definitions.  This difference should be recognised by using different names.  KEY_LEFT is not the same as ACTIONKEY_LEFT so they should NOT use the same variable names.  :)

Here these values represent actions that are performed in response to various remote control button presses.