Network: Possible problems in different DHCP scenarios are fixed
There can be several scenarios when working with local IP.

DHCP grants IP and we also use this IP as Static IP:
In this case there is stability with respect to the code from before, DHCP-router maintains DNS granted by our ISP provider.

First problem that may arise, the scenario changes, the user chooses a static IP that is not granted by DHCP:

So in this case if we keep DHCP-router in the system restart the DNS by default will be set to 0.0.0.0 since that IP was not the one given by DHCP,
as a consequence the DNS is lost.

To solve this we differentiate in the config usage between using DHCP-router and using Static IP or Custom so that this does not happen,
If we choose DHCP to work, it will use any Google assigned provider, etc ...

But if we choose static or custom IP, DHCP-router will not see it in the configuration since this IP may have been granted by DHCP but it may also not have been granted.

The other problem fixed is the indexing of the config chosen in NameServer configuration:

Before we had to press OK to save the new chosen DNS config (resolv.conf is not indexed in iface) and then we had to press OK again for this to happen.
Now this is solved, the new DNS chosen by the user is recorded and indexed with a click of OK.