[config.py] Fix ConfigAction
ConfigAction had a dummy getMulti() method which resulted in the C++ layer not getting the requited text data to display on the user interface.  This was generating "[eListboxPythonConfigContent] second value of tuple is not a tuple." errors in the logs.  The solution was to provide appropriate text to be displayed.

For now the text is an empty string if the entry is not selected and the translatable message "<Press OK to perform action>" if the entry is currently selected.