Don't install PicEmu2 till it gets fixed:
This ugly py file is full of shit no offence to the author but it is how it is.

Unless it gets a proper fix we should avoid installing it.

The result will be not working of emu picons in OctEtFHD-skin and iFlatFHD-skin

The loop bug discovered by IanSav and that's enough proof for me to ignore this file.