[Directories.py] Add fileReadLine() and fileReadLines()
* [Directories.py] Add fileReadLine() and fileReadLines()

The two new methods create a central library for Enigma2 code to read and extract data from disk files.

There is a lot of code that that does a open(filename, "r").read().strip().  This code generally offers no protection against errors and crashes.  These new methods perform a similar function but errors are reported and defaults can easily be assigned if the file operations fail.

If debug mode is enabled the execution stack is used to nominate the calling code module and the line in that module that called these methods.  This should allow debugging when required.  Debugging should not be used by default as this will add significant time to the code execution (while the program stack is processed).

* [Directories.py] Don't report file not found errors

In the new fileReadLine() and fileReadLines() methods don't bother reporting errors if the file does not exist, just return the default value.  If debug mode is enabled you will know the file did not exist as there will be no error message yet the default value will have been used.