InstallWizard.py: add for next step OK key
Now with OK key we can make a return to the list if the box does not have internet.

Before we had to use the EXIT key to make a return to the list repeating step and not parsing the next step in XML.

Now XML is analyzed to step "Check the internet connection again" and we can use  the "OK" key if we find the message "you do not have an internet connection.