FastScan: set default fastscan.auto false
fastscan.auto does a background search from standby mode to the providers involved.

In installations with a rotor, this is not a good option, since if this parameter is enabled and we are on another satellite, 90 seconds after putting the box in standby it will move the antenna to the satellite where the specific fastscan providers are.

When we turn on the box again, it moves the antenna to the satellite where we have the tv start service.

Better that the user decides to enable it.