Introduce BoxInfo to replace boxbranding
BoxInfo is a new Python class in SystemInfo that is designed to replace the old boxbranding system.  The class offers methods to list, get, set or delete system information variables.  The code implements controls to protect variables/values defined to be immutable.

The code is designed to find and use an enigma kernel module to supply the system variables.  If the kernel module is loaded the variables will be directly extracted from the kernel module.  If the module is found but not loaded then the variables will be extracted from the module directly with modinfo.  (This same concept can allow modules from non active slots or partitions to be accessed for purposes like MultiBoot.)

The BoxInfo system also defines and allows a file called "/etc/enigma.conf" to be created and supply replacement / override values for the kernel module.  This feature is designed to allow developers to simulate other boxes or settings to facilitate firmware development and testing.  The file is a simple text format with each line replacing or adding a new kernel module definition.  Each line has the form "variable=value".  Note that spaces *ARE* significant.

When a variable is initially loaded the code attempts to convert text string representations of Python objects into the real Python object.  So, for example, if the string contains '("One", "Two", "Three")' then this will be converted to a tuple with the three string elements.  A similar conversion is done for lists.  The single word "None" in any case is converted to None.  The single words "False", "No", "Off" and "Disabled" in any case are converted to False while "True", "Yes", "On" and "Enabled" in any case are converted to True.  Binary, octal, decimal and hexadecimal numbers are converted to Python integers and floating point number strings are converted to type float.  Please note that at this time only single conversions are performed.  Thus a conversion of a string to a list will not process the elements of the list.

This is an initial presentation of a proposal I designed and presented many months ago.  Given there has been no progress with the concept I am committing this code to further progress and discussion.  This initial variable names have been taken from the OpenVision interpretation of the boxbranding code.  The variable names should be better defined, as suggested / requested by WanWizard on OpenPLi, such changes can easily be made to this code infrastructure.