ActionMap.py: establish sanity in undefined actions
The undefined action list must be eliminated within the range occupied by the action of a key.

From this list, only what the code has been granted in the action of the key should remain and remove it as an action that we do not have defined in the keymap.

We must always do it safely, for this we always do it with a copy.

before we simply append the actions of the undefined list.