Network Nameserver Config: improve the stability of unactivated DNS changes
We make sure not to lose the DNS: this can happen when the name server is changed and the configuration is not activated.

I unlinked addNameserver

getNameserverList returns the length of the string,
If DHCP-router is used and DNS is changed to one of the available servers manually, nameserver can return a string [0, 0, 0, 0] if a DNS change is canceled without activating it.