SoftwareUpdate.py: choose more robust code to update
With this change we do not give the option to Update and reboot (recommended), it changes to Update and ask to reboot (recommended).
some reasons for this change:

Disadvantages of Update and reboot (recommended):

This not check OpkgComponent.CMD_UPGRADE, it would also be necessary to give an execution margin time for this action before executing a restart, also we do not see any information when it is executed, we only view a GUI screen with TryQuitMainloop, retvalue = 4 where only informs user in GUI to be patient that it will take few seconds.direct and dizzying dump to reboot the system reveal that this way of updating is not the correct one

Advantages of Update and ask to reboot (recommended):

Check the packages and we make sure that they are installed, they inform the user live of the count of updated packages and their names, once finished (execution with or without errors depends on whether we have cache errors in the feeds) we have the question if we want to reboot. ... then let's run an improved and more robust update.