Change all AlphaBlend to AlphaTest (use PNGs again)
First of all fuck SVGs! Why should we force users to render a SVG file via GPU just for the sake of something new?! This is hardcore shit!

Second get rid of the shitcode pushed into PLi, I feel sorry for PLi and the situation they're in. I do remember there was a time we called PLi as the most stable image and now we should be scared of upcoming changes as they may fuck up something!