Network.py: don't use a conditional to apply the append iface method
flags collects the state of the config in the adapters data, from the UI, self.configuredInterfaces must always be present and use the append method without any conditions to see the state in the selection of network adapters.

Until now, we did not see the adapter connection status pixmap (green or gray buttons) as an indication of the adapter status, in skin  pixmap default did not show anything in NetworkAdapterSelection.

Now we can see it.

Remember that you must have two types of iface available in the box to be able to see this in the selection of WiFi and Ethernet adapter