setup.xml: add filters in LedDisplay for machine features
We do not have any proc filter to see the config list with these characteristics, this should not be the case since we see non-existent config in certain STBs.
e.g. osmio4kplus.
We saw a list config with LED color, during recording operations, normal or sleep mode, when this does not exist in the type of LedDisplay of this machine (No LED color blue etc..")