[config.py] Fix ConfigDirectory
In a previous update I added the property to activate the getValue() and setValue() methods in this code believing that this was what was intended.  It appears that these methods are incorrect and causing code failures.  This change restores the previous methods from the ConfigText class that were previously in use.