[International.py] Code enhancements
* [International.py] Code enhancements

- Make the getAvailablePackages(), getInstalledPackages() and getPackageDirectories() methods support an update option so that the code can be used as a fetch method to return the values to other modules.  This was done to discourage developers from reaching into this module to look at the internal variables.
- Change the fall back value for setlocale() from "en_US" to "POSIX".  This frees up the image from needing to have the "en_US" full locale settings installed.
- Add the getLanguageTranslated() method to allow the translated name of the language to be accessed as a variable.
- Add a new method, getPurgablePackages(), that can be used to list all installed language/locale packages that are installed but not in the defined list, PERMANENT_LOCALES, of locales that should not be removed in a purge operation.

* [International.py] Make some more corrections

- Correct the getPurgablePackages() method. Use a shallow copy of PERMANENT_LOCALES!
- Add a new method, getPermanentLocales(), to report the locales nominated as permanent.
- Rework runPackageManager() to return a status value as well as a message to address potential language/translation issues.
- Use the phrase "locale/language" more consistently.

* [International.py] Add support for ISO 3166-2 subdivisions

This change allows support for ISO 3166-2 country subdivision codes.  The subdivisions carry over the data from the parent ISO 3166 definitions.  The first such code to be added is "ES-GA" for the Galacia subdivision within Spain.  Thus Galacian will now have the locale code of "gl_ES-GA".

Developers and skinners need to be aware that country codes may now be longer (typically 5 characters) for such subdivisions.  This change also means that all such subdivisions will now require their own flags names in the same style of the subdivision.  In this case the new flag will have the file name "es-ga.png".

* [International.py] Correct getPackageDirectories()

This method and associated variable was miss-named.  The method implied that packages were being listed when in fact it was installed locale and language directories.  This error led to the getPurgablePackages() method not working as intended.  This is now fixed.  Thanks to @Jbleyel for helping diagnose the issue.

Added the localeToPackage() and languageToPackage() methods to make similar data more readily available to other Enigma2 modules.

* Only set en_US as PERMANENT_LOCALES

No other language should be added to this list in OV, only English matters.

Co-authored-by: Persian Prince <persianpros@yahoo.com>