[Information.py] Correct ImageInformation mtd data
Some of the data entries in the mdt area were incorrectly defined.

Note that the source information comes from the build system and *does not* correctly define the partitions actually used.  The issue is the the build system assumes that MultiBoot is not in use.  If MultiBoot *is* used then the default partitions are still reported even if, in reality, other partitions are actually used.